Parameterize naming convention in buildable property tests
Combine BuildableBeanPropertyTest and BuildablePrefixlessPropertyTest into a single parameterized test.
